[fix] cannot get all products

+++ STYLiSH/src/main/java/tw/appworks/school/example/stylish/repository/product/IProductProjection.java
@@ -1,5 +1,7 @@
 package tw.appworks.school.example.stylish.repository.product;
 
+import java.math.BigDecimal;
+
 public interface IProductProjection {
 
     Long getId();
@@ -34,5 +36,5 @@ public interface IProductProjection {
 
     String getColorName();
 
-    Double getRate();
+    BigDecimal getRate();
 }
